add refactor menu
