+++ interactive-data-collection.js
@@ -1,70 +1,8 @@
 const inquirer = require('inquirer');
-const { CollectionUtils } = require('./collectionUtils');
-const newman = require('./run-newman');
 const { Collection } = require('./Collection');
 const { Environment } = require('./Environment');
+const conOp = require('./collectionOperations');
 
-/**
- * Esegue un'operazione interattiva basata sul tipo di dato selezionato.
- *
- * @param {string} dataType - Tipo di operazione da eseguire (Inserisci, Cancella, Modifica, Fine)
- * @param {Collection} collectionData - Dati della collection
- * @param {Environment} environmentData - Dati dell'environment
- * @returns {void} Non ritorna nulla
- */
-async function performOperation(dataType, collectionData, environmentData) {
-  const utils = new CollectionUtils(collectionData, environmentData);
-  switch (dataType) {
-    case 'Mostra':
-      await utils.showInputList();
-      break;
-    case 'Modifica':
-      await utils.modifyValues();
-      break;  
-    case 'Elimina':
-      await utils.removeValues();
-      break;
-    case 'Reset input':
-      await utils.resetValues();
-      break;
-    case 'Aggiungi':
-      await utils.checkSelectedCollection();
-      break;    
-    case 'Esegui':
-      await newman.runNewman(collectionData, environmentData);
-      const { afterExecutionChoice } = await inquirer.prompt([
-        {
-          type: 'list',
-          name: 'afterExecutionChoice',
-          message: 'Cosa vuoi fare dopo l\'esecuzione della collection?',
-          choices: ['Esegui un\'altra collection', 'Riesegui la collection', 'Termina'],
-        },
-      ]);
-      switch (afterExecutionChoice) {
-        case 'Esegui un\'altra collection':
-          // Aggiungi la logica per eseguire un'altra collection
-          // Ad esempio, richiedi il nome della nuova collection e dell'environment
-          // e chiama la funzione interactiveData con i nuovi dati.
-          break;
-        case 'Riesegui la collection':
-          await newman.runNewman(collectionData, environmentData);
-          break;
-        case 'Termina':
-          await utils.exit();
-          break;
-        default:
-          // Se la scelta non corrisponde a nessun caso, non fare nulla 
-          break;
-      }
-      break;
-    case 'Termina':
-      await utils.exit();
-      break;
-    default:
-      // Se il dataType non corrisponde a nessun caso, non fare nulla 
-      break;
-  }
-}
 
 /**
  * Conduce un'interazione interattiva con l'utente per inserire, cancellare o modificare dati dinamicamente.
@@ -87,7 +25,7 @@ async function interactiveData(collectionName, environmentName) {
       },
     ]);
 
-    await performOperation(dataType, collectionData, environmentData);
+    await conOp.performOperation(dataType, collectionData, environmentData);
   }
 }
